% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cloc-package.r
\docType{package}
\name{cloc-package}
\alias{cloc-package}
\title{A package to Count Lines of Code, Comments and Whitespace in Source Files and Archives}
\description{
A package to Count Lines of Code, Comments and Whitespace in Source Files and Archives
}
\author{
Bob Rudis (@hrbrmstr)
}

