% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cloc.R
\name{cloc}
\alias{cloc}
\title{Count lines of code, comments and whitespace in source files/archives}
\usage{
cloc(source, extract_with = NULL)
}
\arguments{
\item{source}{thing to extract from}

\item{extract_with}{thing}
}
\value{
\code{tbl_df}
}
\description{
Count lines of code, comments and whitespace in source files/archives
}

