% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc-by-file.r
\name{cloc_by_file}
\alias{cloc_by_file}
\title{Count lines of code, comments and whitespace in source files/archives by file}
\usage{
cloc_by_file(source, extract_with = NULL)
}
\arguments{
\item{source}{file, directory or archive to read from (can be a valid URL)}

\item{extract_with}{passed into \code{cloc} command line. This option is only
needed if cloc is unable to figure out how to extract the contents of
the input file(s) by itself.}
}
\value{
tibble
}
\description{
Count lines of code, comments and whitespace in source files/archives by file
}
\note{
Unlike it's \code{cloc()} counterpart, no percentages are reported by this function
}
\examples{
# whole dir
str(cloc_by_file(system.file("extdata", package="cloc")))

# single file
str(cloc_by_file(system.file("extdata", "qrencoder.cpp", package="cloc")))
}
