% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cloc.R
\name{cloc_cran}
\alias{cloc_cran}
\title{Count lines of code (etc) from source packages on CRAN}
\usage{
cloc_cran(pkgs, repos = getOption("repos"), contriburl = contrib.url(repos,
  "source"))
}
\arguments{
\item{pkgs}{names of pkgs}

\item{repos}{repos}

\item{contriburl}{}
}
\value{
\code{tbl_df}
}
\description{
Count lines of code (etc) from source packages on CRAN
}

