% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc-git.R
\name{cloc_git}
\alias{cloc_git}
\title{Count lines of code, comments and whitespace in a git tree}
\usage{
cloc_git(repo_dir, commit = ".")
}
\arguments{
\item{repo_dir}{path to git repo}

\item{commit}{"\code{.}" for the current source tree or the commit identifier for a specific commit}
}
\value{
tibble
}
\description{
Count lines of code, comments and whitespace in a git tree
}
\examples{
\dontrun{
cloc_git("~/packages/cloc", "3643cd09d4b951b1b35d32dffe35985dfe7756c4")
}
}
