% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-pkg-src.R
\name{cloc_pkg}
\alias{cloc_pkg}
\title{Count lines of code, comments and whitespace in a package}
\usage{
cloc_pkg(source = ".", extract_with = NULL)
}
\arguments{
\item{source}{file, directory or archive to read from (can be a valid URL)}

\item{extract_with}{passed into \code{cloc} command line. This option is only
needed if cloc is unable to figure out how to extract the contents of
the input file(s) by itself.}
}
\value{
data frame (tibble)
}
\description{
Think of this as \code{\link[=cloc]{cloc()}} with saner defaults for packages. Skips common
IDE tempdirs, \code{.git}, \code{inst}, \code{man}.
}
\examples{
# by dir
cloc(system.file("extdata", package="cloc"))

# by file
cloc(system.file("extdata", "qrencoder.cpp", package="cloc"))

# requires a network connection therefore is set for you to run it manually
\dontrun{
# from a url
cloc("https://rud.is/dl/cloc-1.74.tar.gz")
}
}
