% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clock-pkg-md.R
\name{cloc_pkg_md}
\alias{cloc_pkg_md}
\title{Run \code{\link[=cloc_pkg]{cloc_pkg()}} on active package and format it as a markdown table for knitting into reports}
\usage{
cloc_pkg_md()
}
\description{
Run \code{\link[=cloc_pkg]{cloc_pkg()}} on active package and format it as a markdown table for knitting into reports
}
\note{
Target application is output to Rmd files
}
