% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sloc-rm-comments.R
\name{cloc_remove_comments}
\alias{cloc_remove_comments}
\title{Strip comments and white space from a single source file}
\usage{
cloc_remove_comments(source_file)
}
\arguments{
\item{source_file}{path to source file}
}
\value{
character vector containing only code blocks
}
\description{
Pass in a path to a source file and retrieve a version of the source file
without comments (or white space).
}
\examples{
cloc_remove_comments(system.file("extdata", "qrencoder.cpp", package="cloc"))
}
