% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc-package.r
\docType{package}
\name{cloc-package}
\alias{cloc-package}
\title{Count Lines of Code, Comments and Whitespace in Source Files and Archives}
\description{
Counts blank lines, comment lines, and physical lines of source code in source
files/trees/archives. An R wrapper to the Perl \code{cloc} utility
\url{https://github.com/AlDanial/cloc} by @AlDanial.
}
\author{
Bob Rudis (bob@rud.is)
}
