% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cloc-cran.r
\name{cloc_cran}
\alias{cloc_cran}
\title{Count lines of code (etc) from source packages on CRAN}
\usage{
cloc_cran(pkgs, repos = getOption("repos"),
  contrib_url = utils::contrib.url(repos, "source"),
  .progress = interactive())
}
\arguments{
\item{pkgs}{names of packages}

\item{repos}{character vector, the base URL(s) of the repositories to use,
i.e., the URL of the CRAN master such as "\code{https://cran.r-project.org}"
or its Statlib mirror, "\code{http://lib.stat.cmu.edu/R/CRAN}".}

\item{contrib_url}{URL(s) of the contrib sections of the repositories. Use
this argument only if your repository mirror is incomplete, e.g.,
because you burned only the ‘contrib’ section on a CD. Overrides
argument repos.}

\item{.progress}{show a progress bar? Default: \code{TRUE} if running interactively.}
}
\value{
tibble
}
\description{
Count lines of code (etc) from source packages on CRAN
}
\examples{
# requires a network connection therefore is set for you to run it manually
\dontrun{
cloc_cran(c("archdata", "hrbrthemes", "iptools", "dplyr"))
}
}
